pipewire logo as icon
This adds the PipeWire logo as the icon. The PipeWire logo was created
by Jakub Steiner and is licensed under CC0. More information at
https://commons.wikimedia.org/wiki/File:PipeWire_logo.svg and
https://creativecommons.org/publicdomain/zero/1.0/.

Therefore, it is acceptable to use this icon.

Furthermore, this icon will change to match the color of the other
system tray icons in the notification area depending on the Plasma
theme.